Actually fixed intake issues maybe

+++ 2018CompetitionBot/src/Commands/IntakeUntilLimitCmd.h
@@ -15,8 +15,8 @@ public:
 
 private:
 	bool jawsSetToSpring;
-	int lastMoveTime = 0;
-	int lastAtJaws = 0;
+	double lastMoveTime = 0;
+	double lastAtJaws = 0;
 };
 
 #endif  // IntakeeUntilDistanceCmd_H
